pyproject.toml
[tool.poetry]
name = "drf-apis-in-a-nutshell-darkcloudb"
version = "0.1.0"
description = "REST API"
authors = ["Billy Yip <birii.yip@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
requests = "^2.26.0"
django-cors-headers = "^3.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
